Move Horizon SDK into subfolders
