+++ src/index.ts
@@ -5,27 +5,27 @@
 const version = require("../package.json").version;
 
 // Expose all types
-export * from "./horizon_api";
-export * from "./server_api";
+export * from "./horizon/horizon_api";
+export * from "./horizon/server_api";
 
 // stellar-sdk classes to expose
-export * from "./account_response";
+export * from "./horizon/account_response";
 export * from "./errors";
 export { Config } from "./config";
-export { Server } from "./server";
+export { Server } from "./horizon/server";
 export {
   FederationServer,
   FEDERATION_RESPONSE_MAX_SIZE
-} from "./federation_server";
+} from "./federation/federation_server";
 export {
   StellarTomlResolver,
   STELLAR_TOML_MAX_SIZE
-} from "./stellar_toml_resolver";
+} from "./federation/stellar_toml_resolver";
 export {
   default as HorizonAxiosClient,
   SERVER_TIME_MAP,
   getCurrentServerTime
-} from "./horizon_axios_client";
+} from "./horizon/horizon_axios_client";
 export * from "./utils";
 
 // expose classes and functions from stellar-base
